# schemas/v1/http-api.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EHAI Execution Plane",
    "version": "2"
  },
  "paths": {
...
    "/api/v1/event-consumers": {
      "post": {
        "tags": [
          "event-consumers"
        ],
        "summary": "Register Event Consumer",
        "operationId": "registerEventConsumer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterEventConsumerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "EventConsumer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "queries.schema.json#/$defs/EventConsumerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/event-consumers/{consumer_id}": {
      "get": {
        "tags": [
          "event-consumers"
        ],
        "summary": "Get Event Consumer",
        "operationId": "getEventConsumer",
        "parameters": [
          {
            "name": "consumer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_.-]{0,127}$",
              "title": "Consumer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "EventConsumer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "queries.schema.json#/$defs/EventConsumerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/event-consumers/{consumer_id}/batches": {
      "post": {
        "tags": [
          "event-consumers"
        ],
        "summary": "Read Event Consumer Batch",
        "operationId": "readConsumerEvents",
        "parameters": [
          {
            "name": "consumer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_.-]{0,127}$",
              "title": "Consumer Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadEventConsumerBatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "EventConsumerBatch",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "queries.schema.json#/$defs/EventConsumerBatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/event-consumers/{consumer_id}/ack": {
      "post": {
        "tags": [
          "event-consumers"
        ],
        "summary": "Acknowledge Event Consumer",
        "operationId": "acknowledgeConsumerEvents",
        "parameters": [
          {
            "name": "consumer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_.-]{0,127}$",
              "title": "Consumer Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcknowledgeEventConsumerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "EventConsumerAcknowledgement",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "queries.schema.json#/$defs/EventConsumerAcknowledgementResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/events": {
      "get": {
        "summary": "List Events",
        "operationId": "list_events_api_v1_events_get",
        "parameters": [
          {
            "name": "after_event_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{12}$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "After Event Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Durable Event page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "queries.schema.json#/$defs/EventPageResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AcknowledgeEventConsumerRequest": {
        "properties": {
          "batch_token": {
            "type": "string",
            "pattern": "^[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{12}$",
            "title": "Batch Token"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "batch_token"
        ],
        "title": "AcknowledgeEventConsumerRequest"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReadEventConsumerBatchRequest": {
        "properties": {
          "limit": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 100
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ReadEventConsumerBatchRequest"
      },
      "RegisterEventConsumerRequest": {
        "properties": {
          "consumer_id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.-]{0,127}$",
            "title": "Consumer Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "consumer_id"
        ],
        "title": "RegisterEventConsumerRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}